requirements.txt
flask
langchain
chromadb
sentence-transformers
transformers
torch
wikipedia-api
mysql-connector-python
langchain-community